# Optimize hot code paths