@michaelficarra 2023-02-16
@jviide It seems to work when I remove the `^` from the dependency's semver spec, so we'll just do that. I guess they made a breaking change. Thanks for the pointer though.